Fix const type

+++ Cosette/Engine/Ai/Search/SearchConstants.cs
@@ -12,7 +12,7 @@
         public const int DefaultEvaluationHashTableSize = 32;
 
         public const int NullWindowMinimalDepth = 5;
-        public const byte NullWindowDepthReduction = 3;
+        public const int NullWindowDepthReduction = 3;
 
         public const int LMRMinimalDepth = 2;
         public const int LMRMovesWithoutReduction = 2;
